Rename `status` to `gpustats`
This makes it clearer what is is.
