+++ cmd/groundstation/database/interface.go
@@ -1,6 +1,6 @@
 package database
 
-import "github.com/gpuctl/gpuctl/internal/status"
+import "github.com/gpuctl/gpuctl/internal/gpustats"
 
 // define set of operations on the database that any provider will implement
 type Database interface {
@@ -8,8 +8,8 @@ type Database interface {
 	UpdateLastSeen(host string) error
 
 	// record a new data point for a satellite in the data store
-	AppendDataPoint(packet status.GPUStatusPacket) error
+	AppendDataPoint(packet gpustats.GPUStatusPacket) error
 
 	// get the latest metrics for all approved machines
-	LatestData() ([]status.GPUStatusPacket, error)
+	LatestData() ([]gpustats.GPUStatusPacket, error)
 }
