+++ internal/gpustats/GPUDataSource.go
@@ -1,4 +1,4 @@
-package status
+package gpustats
 
 type GPUDataSource interface {
 	GetGPUStatus() (GPUStatusPacket, error)
