+++ internal/gpustats/_testdata/nvidia/ash03.faultycall
