+++ internal/gpustats/_testdata/nvidia/dump_ash01.json
