+++ internal/gpustats/_testdata/nvidia/dump_ash02.xml
