+++ internal/gpustats/_testdata/nvidia/dump_ash04.xml
