+++ internal/gpustats/_testdata/nvidia/dump_ash05.xml
