+++ internal/gpustats/_testdata/nvidia/dump_gpu01.xml
