+++ internal/gpustats/_testdata/nvidia/dump_gpu02.xml
