+++ internal/gpustats/_testdata/nvidia/dump_gpu03.xml
