+++ internal/gpustats/_testdata/nvidia/dump_gpu04.json
