+++ internal/gpustats/_testdata/nvidia/dump_gpu05.xml
