+++ internal/gpustats/_testdata/nvidia/dump_gpu10.xml
