+++ internal/gpustats/_testdata/nvidia/dump_gpu30.json
