+++ internal/gpustats/_testdata/nvidia/dump_gpu30_altered.json
