+++ internal/gpustats/_testdata/nvidia/dump_gpu36.xml
