+++ internal/gpustats/_testdata/nvidia/dump_texel12.json
