+++ internal/gpustats/_testdata/nvidia/gatherdata.sh
