+++ internal/gpustats/_testdata/nvidia/gpu05.corruptedxml
