+++ internal/gpustats/nvidia.go
@@ -1,4 +1,4 @@
-package status
+package gpustats
 
 import (
 	"encoding/xml"
