+++ internal/gpustats/nvidia_test.go
@@ -1,4 +1,4 @@
-package status
+package gpustats
 
 import (
 	"bytes"
